refactor(painting): Перенес creators в отдельный каталог
