+++ src/main/java/com/example/paint/interactors/creators/OvalCreator.java
@@ -1,4 +1,4 @@
-package com.example.paint.interactors.painting;
+package com.example.paint.interactors.creators;
 
 import com.example.paint.interactors.shapes.Oval;
 import javafx.scene.shape.Ellipse;
