+++ src/main/java/com/example/paint/interactors/creators/ShapeCreator.java
@@ -1,0 +1,7 @@
+package com.example.paint.interactors.creators;
+
+import javafx.scene.shape.Shape;
+
+public abstract class ShapeCreator<MyShape> {
+    public abstract Shape make(MyShape shape);
+}
