+++ src/main/java/com/example/paint/interactors/creators/SquareCreator.java
@@ -1,11 +1,13 @@
-package com.example.paint.interactors.painting;
+package com.example.paint.interactors.creators;
 
-import com.example.paint.interactors.shapes.Rectangle;
+import com.example.paint.interactors.shapes.Square;
+import javafx.scene.shape.Rectangle;
 
-public class RectangleCreator extends ShapeCreator<Rectangle> {
+public class SquareCreator extends ShapeCreator<Square> {
     javafx.scene.shape.Rectangle rectangle = new javafx.scene.shape.Rectangle();
 
-    public javafx.scene.shape.Rectangle make(Rectangle shape) {
+    public Rectangle make(Square shape) {
+        shape.setWidth(shape.getHeight());
         rectangle.setWidth(shape.getWidth());
         rectangle.setHeight(shape.getHeight());
         rectangle.relocate(shape.getStartCoordinate().getX(), shape.getStartCoordinate().getY());
