--- src/main/java/com/example/paint/interactors/painting/ShapeCreator.java
@@ -1,7 +1,0 @@
-package com.example.paint.interactors.painting;
-
-import javafx.scene.shape.Shape;
-
-public abstract class ShapeCreator<MyShape> {
-    abstract Shape make(MyShape shape);
-}
